pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stindex"
version = "0.1.0"
description = "Spatiotemporal Index Extraction from Unstructured Text"
readme = "README.md"
authors = [
    {name = "STIndex Team", email = "stindex@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Linguistic",
]
keywords = ["nlp", "spatiotemporal", "information-extraction", "ner", "geocoding", "temporal-extraction"]
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "pydantic>=2.0.0",
    "spacy>=3.7.0",
    "dateparser>=1.2.0",
    "pendulum>=3.0.0",
    "geopy>=2.4.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
transformers = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/stindex"
Documentation = "https://stindex.readthedocs.io"
Repository = "https://github.com/yourusername/stindex"
"Bug Tracker" = "https://github.com/yourusername/stindex/issues"

[project.scripts]
stindex = "stindex.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stindex*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=stindex --cov-report=term-missing"
